requirements.txt
flask==3.0.2
numpy==1.26.4
pandas==2.2.1
nltk==3.8.1
python-dotenv==1.0.1
pytest==8.0.2
chromadb==0.4.24
langchain==0.1.12
pdfplumber==0.10.3
sentence-transformers==2.5.1
groq==0.4.2 